trainsim: move salut à tous to libtsim
