+++ train-sim/libtsim/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "libtsim"
+version = "3.0.0"
+edition = "2024"
+license = "LGPL-3.0"
+
+[dependencies]
