+++ train-sim/libtsim/src/lib.rs
@@ -1,0 +1,8 @@
+// TODO add lint for missing docs
+
+pub use crate::range::Range;
+pub use crate::range::RangeKind;
+pub use crate::range::RangeMap;
+
+mod integration;
+mod range;
